refactor normalized state/tools
