+++ src/store/normalized/normalizer.ts
@@ -1,0 +1,65 @@
+import { schema as nSchema, denormalize, normalize, Schema } from 'normalizr';
+import { RootState } from '..';
+
+export class Normalizer {
+  private _schema: nSchema.Entity;
+  private _listSchema: Schema;
+
+  constructor(schema: nSchema.Entity) {
+    this._schema = schema;
+    this._listSchema = [schema];
+  }
+
+  public normalize = (item) => {
+    if (Array.isArray(item)) {
+      return this.normalizeMany(item);
+    }
+
+    return this.normalizeSingle(item);
+  };
+
+  public denormalize = (idOrIds: string | string[], state: RootState) => {
+    if (Array.isArray(idOrIds)) {
+      return this.denormalizeMany(idOrIds, state);
+    }
+
+    return this.denormalizeSingle(idOrIds, state);
+  };
+
+  private normalizeMany(items) {
+    this.throwIfInvalid(items);
+
+    return normalize(items, this._listSchema);
+  }
+
+  private normalizeSingle(item) {
+    this.throwIfInvalid([item]);
+
+    return normalize(item, this._schema);
+  }
+
+  private denormalizeSingle(id: string, state: RootState) {
+    const denormalized = denormalize(id, this._schema, state.normalized);
+
+    if (denormalized) {
+      denormalized.__denormalized = true;
+    }
+
+    return denormalized;
+  }
+
+  private denormalizeMany(ids: string[], state: RootState) {
+    const denormalizedList = denormalize(ids, this._listSchema, state.normalized);
+    return denormalizedList.map((denormalized) => ({ ...(denormalized as any), __denormalized: true }));
+  }
+
+  private throwIfInvalid(items) {
+    items.forEach((item) => {
+      if (item.__denormalized) {
+        throw new Error(
+          'Tried to normalize an object that was previously denormalized from the store. This can cause infinite loops.'
+        );
+      }
+    });
+  }
+}
